+++ changelog.md
@@ -4,6 +4,6 @@
 
 * Maintain multiple snapshots of a document
 * Simple path query for archive elements
-* **TODO**: Retrieve document snapshots from archive
+* Retrieve document snapshots from archive
 * **TODO**: Maintain archives on disk as Json or Yaml files
 * **TODO**: PyPI package